refactor: rename applyCards to applyCard
