+++ functions/src/handlers/applyCard/index.ts
